ANDB parser integration into native code. All changes since the broken build are back now. Refactored most of the code to avoid messy header file order dependecies.
