+++ server/experimental/src/common/platform.h
@@ -5,6 +5,7 @@
 // these differences should be thin wrappers fir for encapsulation, major
 // differences don't belong here.
 
+namespace andb {
 #ifdef _MSC_VER
 // compare two C strings in case insensitive mode
 inline int andb_strcmp_nocase (const char* str1, const char* str2)
@@ -18,3 +19,4 @@ inline int andb_strcmp_nocase(const char* str1, const char* str2)
     return strcasecmp (str1, str2);
 }
 #endif
+}
